requirements.txt.txt
streamlit==1.35.0
pandas==2.2.2
snowflake-connector-python==3.10.1
google-generativeai==0.4.1